chore: adding use hooks on comments services
